requirements.txt
# 核心依赖
requests>=2.31.0
Pillow>=10.0.0

# 浏览器自动化
playwright>=1.40.0
playwright-stealth>=0.0.6

# GUI 框架
PySide6>=6.6.0

# 数据处理
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Excel导出
openpyxl>=3.1.0

# 其他工具
python-dotenv>=1.0.0
